Lower @differentiable to SIL [differentiable] attribute.
PiperOrigin-RevId: 186193136

+++ include/swift/AST/Attr.h
@@ -1306,7 +1306,7 @@ public:
 class DifferentiableAttr : public DeclAttribute {
 private:
   /// The number of arguments specified in 'withRespectTo:'.
-  size_t NumArguments;
+  unsigned NumArguments;
   /// The name of the function which the declared function differentiates.
   DeclName GradFuncName;
   DeclNameLoc GradFuncNameLoc;
